requirements.txt
image==1.5.33
piexif==1.1.3
Pillow~=9.4.0